backtester/requirements.txt
uvicorn==0.23.2
fastapi==0.101.1
websockets==11.0.3
pandas==2.0.3
python-env-loader==0.0.2
